refactor: improve disruptor (DIS-33)
The disruptor was a bit too complex. Iterations over multiple DOM Nodes were made but only one element was displayed in the end.

Now the disruptor uses a copy template to show the markup in the lightbox. With the custom lightbox close Event the disruptor message can be shown at the top of the page.